// src/modules/setting/interface/setting.interface.ts
export interface IStetting {
    sitePhone: string;
    address: string;
    tgLink: string;
    paymentOperator: number;
    penaltyOperator: number;
    eskiz_token?: string;
    logo?: string;
    instaLink: string;
    fbLink: string;
    ad_price?: number;
    delivery_price?: number;
    marketingGroup?: string;
    supportBot?: string;
    supportPhone?: string;
    highestPriceForCheapProducts?: number;
}